refactor(Events): Move from `pubsub` module to `events` crate
