+++ node/Cargo.toml
@@ -12,6 +12,7 @@ crate-type = ["cdylib"]
 
 [dependencies]
 stencila = { path = "../rust/stencila" }
+events = { path = "../rust/events" }
 formats = { path = "../rust/formats" }
 graph = { path = "../rust/graph" }
 neon = { version = "=0.9.1", default-features = false, features = ["napi-6", "event-queue-api"] }
