+++ rust/stencila/src/files.rs
@@ -1,5 +1,6 @@
-use crate::{pubsub::publish, utils::schemas};
+use crate::utils::schemas;
 use defaults::Defaults;
+use events::publish;
 use formats::{Format, FORMATS};
 use schemars::{schema::Schema, JsonSchema};
 use serde::Serialize;
