+++ rust/stencila/src/projects.rs
@@ -4,9 +4,9 @@ use crate::documents::DOCUMENTS;
 use crate::errors::attempt;
 use crate::files::{File, FileEvent, Files};
 use crate::methods::import::import;
-use crate::pubsub::publish;
 use crate::sources::{self, Source, SourceDestination, SourceTrait};
 use crate::utils::schemas;
+use events::publish;
 use eyre::{bail, Result};
 use graph::{Graph, GraphEvent, GraphEventType};
 use graph_triples::{resources, Resource};
